feat: Split permissions server and client code
