+++ app/routes/admin+/cache.tsx
@@ -20,7 +20,7 @@ import {
   searchCacheKeys,
 } from "#app/utils/cache.server.ts";
 import { useDebounce, useDoubleCheck } from "#app/utils/misc.tsx";
-import { requireUserWithRole } from "#app/utils/permissions.ts";
+import { requireUserWithRole } from "#app/utils/permissions.server.ts";
 
 import type { ActionFunctionArgs, LoaderFunctionArgs } from "@remix-run/node";
