+++ app/routes/users+/$username_+/notes.$noteId.tsx
@@ -15,12 +15,9 @@ import { StatusButton } from "#app/components/ui/status-button.tsx";
 import { requireUserId } from "#app/utils/auth.server.ts";
 import { prisma } from "#app/utils/db/db.server.ts";
 import { getNoteImgSrc, useIsPending } from "#app/utils/misc.tsx";
-import {
-  requireUserWithPermission,
-  userHasPermission,
-} from "#app/utils/permissions.ts";
+import { requireUserWithPermission } from "#app/utils/permissions.server.ts.ts";
 import { redirectWithToast } from "#app/utils/toast.server.ts";
-import { useOptionalUser } from "#app/utils/user.ts";
+import { useOptionalUser, userHasPermission } from "#app/utils/user.ts";
 
 import type { ActionFunctionArgs, LoaderFunctionArgs } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/react";
