+++ app/utils/permissions.server.ts.ts
@@ -2,8 +2,9 @@ import { json } from "@remix-run/node";
 
 import { requireUserId } from "./auth.server.ts";
 import { prisma } from "./db/db.server.ts";
+import { parsePermissionString } from "./user.ts";
 
-import type { useUser } from "./user.ts";
+import type { PermissionString } from "./user.ts";
 
 export async function requireUserWithPermission(
   request: Request,
@@ -59,49 +60,4 @@ export async function requireUserWithRole(request: Request, name: string) {
     );
   }
   return user.id;
-}
-
-type Action = "create" | "read" | "update" | "delete";
-type Entity = "user" | "note";
-type Access = "own" | "any" | "own,any" | "any,own";
-type PermissionString = `${Action}:${Entity}` | `${Action}:${Entity}:${Access}`;
-function parsePermissionString(permissionString: PermissionString) {
-  const [action, entity, access] = permissionString.split(":") as [
-    Action,
-    Entity,
-    Access | undefined,
-  ];
-  return {
-    action,
-    entity,
-    access: access ? (access.split(",") as Access[]) : undefined,
-  };
-}
-
-export function userHasPermission(
-  user: Pick<ReturnType<typeof useUser>, "roles"> | null | undefined,
-  permission: PermissionString,
-) {
-  if (!user) {
-    return false;
-  }
-  const { action, entity, access } = parsePermissionString(permission);
-  return user.roles.some((role) =>
-    role.permissions.some(
-      (permission) =>
-        permission.entity === entity &&
-        permission.action === action &&
-        (!access || access.includes(permission.access)),
-    ),
-  );
-}
-
-export function userHasRole(
-  user: Pick<ReturnType<typeof useUser>, "roles"> | null,
-  role: string,
-) {
-  if (!user) {
-    return false;
-  }
-  return user.roles.some((r) => r.name === role);
 }
